requirements.txt
pandas
sqlalchemy
psycopg2-binary
pyyaml
confuse
python-aqi